requirements.txt
openai
faiss-cpu
numpy
pdfminer.six
python-dotenv